+++ client/src/__tests__/App.test.jsx
@@ -8,7 +8,7 @@ const mockApp = () => {
   render(
     <BrowserRouter>
       <App />
-    </BrowserRouter>
+    </BrowserRouter>,
   );
 };
 
@@ -19,16 +19,12 @@ beforeEach(() => {
 // runs a cleanup after each test case (e.g. clearing jsdom)
 describe("renders all page elements", () => {
   it("should render nav bar and footer with logo link", () => {
-    const linkElements = screen.getAllByRole("link", {
-      name: /.com/i,
-    });
-
-    linkElements.forEach((element) => expect(element).toBeInTheDocument());
-    expect(linkElements).toHaveLength(2);
+    const linkElement = screen.getByTestId(/logo-naviation-link/i);
+    expect(linkElement).toBeInTheDocument();
   });
 
-  it("should render footer component with copyright info", () => {
-    const footerElement = screen.getByText(/2023 | FreeTypingCamp/i);
+  it("should render footer component with copyright info", async () => {
+    const footerElement = await screen.findByText(/2023/i);
     expect(footerElement).toBeInTheDocument();
   });
 });
